requirements.txt
google-api-python-client
python-dotenv
PyQt6
pyqtdarktheme
isodate
